requirements.txt
langchain
langchain-community
langchain_huggingface
langchain-text-splitters
langchain-experimental
ipykernel
protobuf==3.20.3
python-dotenv
pypdf
pymupdf
bs4
arxiv
wikipedia
chromadb
sentence_transformers
faiss-cpu
langchain-cli
langchain_chroma
langchain-groq
langchain-astradb
langchain-elasticsearch
langchain-mongodb
langchain-ollama
langchain-pinecone
langchain-postgres
langchain-redis
langchain-unstructured
langserve
streamlit
fastapi
pydantic
uvicorn
pandas
duckdb
duckduckgo_search==5.3.1b1
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
pinecone-client
pinecone-text
pinecone-notebooks
torch
torchvision
torchaudio
aws
neo4j==5.14
openai-whisper
tiktoken
numpy
pdf2image
boto3
psycopg2
diffusers 
transformers 
accelerate 
scikit-learn
bark
django
scipy
playsound
sounddevice
pvporcupine
pyaudio 
langgraph
cassio
Authlib
streamlit_authenticator
google-auth 
google-auth-oauthlib 
google-auth-httplib2
llama-cpp-python
faster-whisper
pyttsx3
redis
pytesseract
langchain-redis
